+++ src/d4c.cpp
@@ -19,7 +19,7 @@ namespace {
 // SetParametersForGetWindowedWaveform()
 //-----------------------------------------------------------------------------
 static void SetParametersForGetWindowedWaveform(int half_window_length,
-    int x_length, double current_position, int fs, double current_f0,
+    size_t x_length, double current_position, unsigned int fs, double current_f0,
     int window_type, double window_length_ratio, int *base_index,
     int *safe_index, double *window) {
   for (int i = -half_window_length; i <= half_window_length; ++i)
@@ -49,7 +49,7 @@ static void SetParametersForGetWindowedWaveform(int half_window_length,
 // GetWindowedWaveform() windows the waveform by F0-adaptive window
 // In the variable window_type, 1: hanning, 2: blackman
 //-----------------------------------------------------------------------------
-static void GetWindowedWaveform(const double *x, int x_length, int fs,
+static void GetWindowedWaveform(const double *x, size_t x_length, unsigned int fs,
     double current_f0, double current_position, int window_type,
     double window_length_ratio, double *waveform) {
   int half_window_length =
@@ -87,10 +87,10 @@ static void GetWindowedWaveform(const double *x, int x_length, int fs,
 // GetCentroid() calculates the energy centroid (see the book, time-frequency
 // analysis written by L. Cohen).
 //-----------------------------------------------------------------------------
-static void GetCentroid(const double *x, int x_length, int fs,
-    double current_f0, int fft_size, double current_position,
+static void GetCentroid(const double *x, size_t x_length, unsigned int fs,
+    double current_f0, unsigned int fft_size, double current_position,
     const ForwardRealFFT *forward_real_fft, double *centroid) {
-  for (int i = 0; i < fft_size; ++i) forward_real_fft->waveform[i] = 0.0;
+  for (unsigned int i = 0; i < fft_size; ++i) forward_real_fft->waveform[i] = 0.0;
   GetWindowedWaveform(x, x_length, fs, current_f0,
       current_position, world::kBlackman, 4.0, forward_real_fft->waveform);
   double power = 0.0;
@@ -102,15 +102,15 @@ static void GetCentroid(const double *x, int x_length, int fs,
   fft_execute(forward_real_fft->forward_fft);
   double *tmp_real = new double[fft_size / 2 + 1];
   double *tmp_imag = new double[fft_size / 2 + 1];
-  for (int i = 0; i <= fft_size / 2; ++i) {
+  for (unsigned int i = 0; i <= fft_size / 2; ++i) {
     tmp_real[i] = forward_real_fft->spectrum[i][0];
     tmp_imag[i] = forward_real_fft->spectrum[i][1];
   }
 
-  for (int i = 0; i < fft_size; ++i)
+  for (unsigned int i = 0; i < fft_size; ++i)
     forward_real_fft->waveform[i] *= i + 1.0;
   fft_execute(forward_real_fft->forward_fft);
-  for (int i = 0; i <= fft_size / 2; ++i)
+  for (unsigned int i = 0; i <= fft_size / 2; ++i)
     centroid[i] = forward_real_fft->spectrum[i][0] * tmp_real[i] +
       tmp_imag[i] * forward_real_fft->spectrum[i][1];
 
@@ -122,8 +122,8 @@ static void GetCentroid(const double *x, int x_length, int fs,
 // GetStaticCentroid() calculates the temporally static energy centroid.
 // Basic idea was proposed by H. Kawahara.
 //-----------------------------------------------------------------------------
-static void GetStaticCentroid(const double *x, int x_length, int fs,
-    double current_f0, int fft_size, double current_position,
+static void GetStaticCentroid(const double *x, size_t x_length, unsigned int fs,
+    double current_f0, unsigned int fft_size, double current_position,
     const ForwardRealFFT *forward_real_fft, double *static_centroid) {
   double *centroid1 = new double[fft_size / 2 + 1];
   double *centroid2 = new double[fft_size / 2 + 1];
@@ -133,7 +133,7 @@ static void GetStaticCentroid(const double *x, int x_length, int fs,
   GetCentroid(x, x_length, fs, current_f0, fft_size,
       current_position + 0.25 / current_f0, forward_real_fft, centroid2);
 
-  for (int i = 0; i <= fft_size / 2; ++i)
+  for (unsigned int i = 0; i <= fft_size / 2; ++i)
     static_centroid[i] = centroid1[i] + centroid2[i];
 
   DCCorrection(static_centroid, current_f0, fs, fft_size, static_centroid);
@@ -145,15 +145,15 @@ static void GetStaticCentroid(const double *x, int x_length, int fs,
 // GetSmoothedPowerSpectrum() calculates the smoothed power spectrum.
 // The parameters used for smoothing are optimized in davance.
 //-----------------------------------------------------------------------------
-static void GetSmoothedPowerSpectrum(const double *x, int x_length, int fs,
-    double current_f0, int fft_size, double current_position,
+static void GetSmoothedPowerSpectrum(const double *x, size_t x_length, unsigned int fs,
+    double current_f0, unsigned int fft_size, double current_position,
     const ForwardRealFFT *forward_real_fft, double *smoothed_power_spectrum) {
-  for (int i = 0; i < fft_size; ++i) forward_real_fft->waveform[i] = 0.0;
+  for (unsigned int i = 0; i < fft_size; ++i) forward_real_fft->waveform[i] = 0.0;
   GetWindowedWaveform(x, x_length, fs, current_f0,
       current_position, world::kHanning, 4.0, forward_real_fft->waveform);
 
   fft_execute(forward_real_fft->forward_fft);
-  for (int i = 0; i <= fft_size / 2; ++i)
+  for (unsigned int i = 0; i <= fft_size / 2; ++i)
     smoothed_power_spectrum[i] =
       forward_real_fft->spectrum[i][0] * forward_real_fft->spectrum[i][0] +
       forward_real_fft->spectrum[i][1] * forward_real_fft->spectrum[i][1];
@@ -168,9 +168,9 @@ static void GetSmoothedPowerSpectrum(const double *x, int x_length, int fs,
 // This is the fundamental parameter in D4C.
 //-----------------------------------------------------------------------------
 static void GetStaticGroupDelay(const double *static_centroid,
-    const double *smoothed_power_spectrum, int fs, double f0,
-    int fft_size, double *static_group_delay) {
-  for (int i = 0; i <= fft_size / 2; ++i)
+    const double *smoothed_power_spectrum, unsigned int fs, double f0,
+    unsigned int fft_size, double *static_group_delay) {
+  for (unsigned int i = 0; i <= fft_size / 2; ++i)
     static_group_delay[i] = static_centroid[i] / smoothed_power_spectrum[i];
   LinearSmoothing(static_group_delay, f0 / 2.0, fs, fft_size,
       static_group_delay);
@@ -179,7 +179,7 @@ static void GetStaticGroupDelay(const double *static_centroid,
   LinearSmoothing(static_group_delay, f0, fs, fft_size,
       smoothed_group_delay);
 
-  for (int i = 0; i <= fft_size / 2; ++i)
+  for (unsigned int i = 0; i <= fft_size / 2; ++i)
     static_group_delay[i] -= smoothed_group_delay[i];
 
   delete[] smoothed_group_delay;
@@ -189,31 +189,31 @@ static void GetStaticGroupDelay(const double *static_centroid,
 // GetCoarseAperiodicity() calculates the aperiodicity in multiples of 3 kHz.
 // The upper limit is given based on the sampling frequency.
 //-----------------------------------------------------------------------------
-static void GetCoarseAperiodicity(const double *static_group_delay, int fs,
-    int fft_size, int number_of_aperiodicities, const double *window,
-    int window_length, const ForwardRealFFT *forward_real_fft,
+static void GetCoarseAperiodicity(const double *static_group_delay, unsigned int fs,
+    unsigned int fft_size, int number_of_aperiodicities, const double *window,
+    unsigned int window_length, const ForwardRealFFT *forward_real_fft,
     double *coarse_aperiodicity) {
-  int boundary =
+  unsigned int boundary =
     matlab_round(fft_size * 8.0 / window_length);
-  int half_window_length = window_length / 2;
+  unsigned int half_window_length = window_length / 2;
 
-  for (int i = 0; i < fft_size; ++i) forward_real_fft->waveform[i] = 0.0;
+  for (unsigned int i = 0; i < fft_size; ++i) forward_real_fft->waveform[i] = 0.0;
 
   double *power_spectrum = new double[fft_size / 2 + 1];
   int center;
   for (int i = 0; i < number_of_aperiodicities; ++i) {
     center =
       static_cast<int>(world::kFrequencyInterval * (i + 1) * fft_size / fs);
-    for (int j = 0; j <= half_window_length * 2; ++j)
+    for (unsigned int j = 0; j <= half_window_length * 2; ++j)
       forward_real_fft->waveform[j] =
         static_group_delay[center - half_window_length + j] * window[j];
     fft_execute(forward_real_fft->forward_fft);
-    for (int j = 0 ; j <= fft_size / 2; ++j)
+    for (unsigned int j = 0 ; j <= fft_size / 2; ++j)
       power_spectrum[j] =
         forward_real_fft->spectrum[j][0] * forward_real_fft->spectrum[j][0] +
         forward_real_fft->spectrum[j][1] * forward_real_fft->spectrum[j][1];
     std::sort(power_spectrum, power_spectrum + fft_size / 2 + 1);
-    for (int j = 1 ; j <= fft_size / 2; ++j)
+    for (unsigned int j = 1 ; j <= fft_size / 2; ++j)
       power_spectrum[j] += power_spectrum[j - 1];
     coarse_aperiodicity[i] =
       10 * log10(power_spectrum[fft_size / 2 - boundary - 1] /
@@ -222,26 +222,26 @@ static void GetCoarseAperiodicity(const double *static_group_delay, int fs,
   delete[] power_spectrum;
 }
 
-static double D4CLoveTrainSub(const double *x, int fs, int x_length,
-    double current_f0, double current_position, int f0_length, int fft_size,
-    int boundary0, int boundary1, int boundary2,
+static double D4CLoveTrainSub(const double *x, unsigned int fs, size_t x_length,
+    double current_f0, double current_position, size_t f0_length, unsigned int fft_size,
+    unsigned int boundary0, unsigned int boundary1, unsigned int boundary2,
     ForwardRealFFT *forward_real_fft) {
   double *power_spectrum = new double[fft_size];
 
-  int window_length = matlab_round(1.5 * fs / current_f0) * 2 + 1;
+  unsigned int window_length = matlab_round(1.5 * fs / current_f0) * 2 + 1;
   GetWindowedWaveform(x, x_length, fs, current_f0, current_position,
     world::kBlackman, 3.0, forward_real_fft->waveform);
 
-  for (int i = window_length; i < fft_size; ++i)
+  for (unsigned int i = window_length; i < fft_size; ++i)
     forward_real_fft->waveform[i] = 0.0;
   fft_execute(forward_real_fft->forward_fft);
 
-  for (int i = 0; i <= boundary0; ++i) power_spectrum[i] = 0.0;
-  for (int i = boundary0 + 1; i < fft_size / 2 + 1; ++i)
+  for (unsigned int i = 0; i <= boundary0; ++i) power_spectrum[i] = 0.0;
+  for (unsigned int i = boundary0 + 1; i < fft_size / 2 + 1; ++i)
     power_spectrum[i] =
     forward_real_fft->spectrum[i][0] * forward_real_fft->spectrum[i][0] +
     forward_real_fft->spectrum[i][1] * forward_real_fft->spectrum[i][1];
-  for (int i = boundary0; i <= boundary2; ++i)
+  for (unsigned int i = boundary0; i <= boundary2; ++i)
     power_spectrum[i] += +power_spectrum[i - 1];
 
   double aperiodicity0 = power_spectrum[boundary1] / power_spectrum[boundary2];
@@ -255,20 +255,20 @@ static double D4CLoveTrainSub(const double *x, int fs, int x_length,
 // very high value as the safeguard.
 // If it was voiced section, the aperiodicity of 0 Hz is set to -60 dB.
 //-----------------------------------------------------------------------------
-static void D4CLoveTrain(const double *x, int fs, int x_length,
-    const double *f0, int f0_length, double frame_period,
+static void D4CLoveTrain(const double *x, unsigned int fs, size_t x_length,
+    const double *f0, size_t f0_length, double frame_period,
     double *aperiodicity0) {
   double lowest_f0 = 40.0;
-  int fft_size = static_cast<int>(pow(2.0, 1.0 +
-    static_cast<int>(log(3.0 * fs / lowest_f0 + 1) / world::kLog2)));
+  unsigned int fft_size = static_cast<unsigned int>(pow(2.0, 1.0 +
+    static_cast<unsigned int>(log(3.0 * fs / lowest_f0 + 1) / world::kLog2)));
   ForwardRealFFT forward_real_fft = { 0 };
   InitializeForwardRealFFT(fft_size, &forward_real_fft);
 
   // Cumulative powers at 100, 4000, 7900 Hz are used for VUV identification.
-  int boundary0 = static_cast<int>(ceil(100.0 * fft_size / fs));
-  int boundary1 = static_cast<int>(ceil(4000.0 * fft_size / fs));
-  int boundary2 = static_cast<int>(ceil(7900.0 * fft_size / fs));
-  for (int i = 0; i < f0_length; ++i) {
+  unsigned int boundary0 = static_cast<unsigned int>(ceil(100.0 * fft_size / fs));
+  unsigned int boundary1 = static_cast<unsigned int>(ceil(4000.0 * fft_size / fs));
+  unsigned int boundary2 = static_cast<unsigned int>(ceil(7900.0 * fft_size / fs));
+  for (size_t i = 0; i < f0_length; ++i) {
     if (f0[i] == 0.0) {
       aperiodicity0[i] = 0.0;
       continue;
@@ -287,9 +287,9 @@ static void D4CLoveTrain(const double *x, int fs, int x_length,
 // Caution:
 //   forward_fft is allocated in advance to speed up the processing.
 //-----------------------------------------------------------------------------
-static void D4CGeneralBody(const double *x, int x_length, int fs,
-    double current_f0, int fft_size, double current_position,
-    int number_of_aperiodicities, const double *window, int window_length,
+static void D4CGeneralBody(const double *x, size_t x_length, unsigned int fs,
+    double current_f0, unsigned int fft_size, double current_position,
+    int number_of_aperiodicities, const double *window, unsigned int window_length,
     const ForwardRealFFT *forward_real_fft, double *coarse_aperiodicity) {
   double *static_centroid = new double[fft_size / 2 + 1];
   double *smoothed_power_spectrum = new double[fft_size / 2 + 1];
@@ -315,33 +315,33 @@ static void D4CGeneralBody(const double *x, int x_length, int fs,
   delete[] static_group_delay;
 }
 
-static void InitializeAperiodicity(int f0_length, int fft_size,
+static void InitializeAperiodicity(size_t f0_length, unsigned int fft_size,
     double *const *aperiodicity) {
-  for (int i = 0; i < f0_length; ++i)
-    for (int j = 0; j < fft_size / 2 + 1; ++j)
+  for (size_t i = 0; i < f0_length; ++i)
+    for (unsigned int j = 0; j < fft_size / 2 + 1; ++j)
       aperiodicity[i][j] = 1.0 - world::kMySafeGuardMinimum;
 }
 
 static void GetAperiodicity(const double *coarse_frequency_axis,
     const double *coarse_aperiodicity, int number_of_aperiodicities,
-    const double *frequency_axis, int fft_size, double *aperiodicity) {
+    const double *frequency_axis, unsigned int fft_size, double *aperiodicity) {
   interp1(coarse_frequency_axis, coarse_aperiodicity,
     number_of_aperiodicities + 2, frequency_axis, fft_size / 2 + 1,
     aperiodicity);
-  for (int i = 0; i <= fft_size / 2; ++i)
+  for (unsigned int i = 0; i <= fft_size / 2; ++i)
     aperiodicity[i] = pow(10.0, aperiodicity[i] / 20.0);
 }
 
 }  // namespace
 
-void D4C(const double *x, int x_length, int fs,
-    const double *f0, int f0_length, double frame_period,
-    int fft_size, const D4COption *option, double *const *aperiodicity) {
+void D4C(const double *x, size_t x_length, unsigned int fs,
+    const double *f0, size_t f0_length, double frame_period,
+    unsigned int fft_size, const D4COption *option, double *const *aperiodicity) {
   randn_reseed();
 
   InitializeAperiodicity(f0_length, fft_size, aperiodicity);
 
-  int fft_size_d4c = static_cast<int>(pow(2.0, 1.0 +
+  unsigned int fft_size_d4c = static_cast<int>(pow(2.0, 1.0 +
     static_cast<int>(log(4.0 * fs / world::kFloorF0D4C + 1) /
       world::kLog2)));
 
@@ -353,8 +353,8 @@ void D4C(const double *x, int x_length, int fs,
       world::kFrequencyInterval) / world::kFrequencyInterval);
   // Since the window function is common in D4CGeneralBody(),
   // it is designed here to speed up.
-  int window_length =
-    static_cast<int>(world::kFrequencyInterval * fft_size_d4c / fs) * 2 + 1;
+  unsigned int window_length =
+    static_cast<unsigned int>(world::kFrequencyInterval * fft_size_d4c / fs) * 2 + 1;
   double *window =  new double[window_length];
   NuttallWindow(window_length, window);
 
@@ -373,10 +373,10 @@ void D4C(const double *x, int x_length, int fs,
   coarse_frequency_axis[number_of_aperiodicities + 1] = fs / 2.0;
 
   double *frequency_axis = new double[fft_size / 2 + 1];
-  for (int i = 0; i <= fft_size / 2; ++i)
+  for (unsigned int i = 0; i <= fft_size / 2; ++i)
     frequency_axis[i] = static_cast<double>(i) * fs / fft_size;
 
-  for (int i = 0; i < f0_length; ++i) {
+  for (size_t i = 0; i < f0_length; ++i) {
     if (f0[i] == 0 || aperiodicity0[i] <= option->threshold) continue;
     D4CGeneralBody(x, x_length, fs, MyMaxDouble(world::kFloorF0D4C, f0[i]),
         fft_size_d4c, frame_period * i / 1000.0, number_of_aperiodicities, window,
